+++ modules/task_1/vorobev_l_mid/mid.h
@@ -3,10 +3,11 @@
 #define MODULES_TASK_1_VOROBEV_L_MID_H_
 
 #include <vector>
+#include <string>
 
 std::vector<int> randvector(int size);
-double avgvector(std::vector<int> vec, int sizevec);
-double avgvectorpar(std::vector<int> vec, int sizevec);
+int avgvector(std::vector<int> vec, int sizevec);
+int avgvectorpar(std::vector<int> vec, int sizevec);
 
 
 #endif  // MODULES_TASK_1_VOROBEV_L_MID_H_
